+++ psysh.php
@@ -8,7 +8,8 @@ class PsyshPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized' => ['onPluginsInitialized', 0]
+            'onPluginsInitialized' => ['onPluginsInitialized', 0],
+            'onTwigExtensions' => ['onTwigExtensions', 0],
         ];
     }
 
@@ -17,9 +18,15 @@ class PsyshPlugin extends Plugin
      */
     public function onPluginsInitialized()
     {
+        require_once(__DIR__ . '/global/function.php');
         $this->require_binary();
     }
 
+    public function onTwigExtensions()
+    {
+        require_once(__DIR__ . '/twig/PsyshTwigExtension.php');
+        $this->grav['twig']->twig->addExtension(new PsyshTwigExtension());
+    }
 
     /**
      * PLUGIN METHODS
@@ -35,5 +42,24 @@ class PsyshPlugin extends Plugin
         ob_start(); // Fix https://git.io/vSlBs
         require($psysh_binary);
         ob_end_clean();
+    }
+
+    static function fire_shell_from_twig()
+    {
+        $grav = \Grav\Common\Grav::instance();
+        if(!$grav) return;
+        if(!isset($grav['config'])) return;
+        $psysh_enabled = $grav['config']->get('plugins.psysh.enabled');
+        if(!$psysh_enabled) return;
+
+        /**
+         * This comment will be visible with `whereami`
+         *
+         * =============== WARNING ===============
+         * Shell context is
+         * \Grav\Plugin\PsyshPlugin::fire_shell()
+         * Use $grav['twig'] to debug
+         **/
+        extract(\Psy\Shell::debug(get_defined_vars()));
     }
 }