+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#new)
     * Can now fire shell from twig
+1. [](#improved)
+    * Added help comment for `whereami`
 
 # v1.0.5
 ## 05/04/2017